Tests/LibWeb/Text/expected/content-blocker-shadow-cache-survives-class-refresh.txt

rgb(0, 0, 0)
inline
rgb(1, 2, 3)
none
hasInvalidationRuleCacheBuilds: 1
